+++ src/Table/Filters/Presets/TitleFilter.php
@@ -6,12 +6,12 @@ use Thinktomorrow\Chief\Table\Filters\SearchFilter;
 
 class TitleFilter extends SearchFilter
 {
-    public static function makeDefault(array $columns = [], array $dynamicKeys = ['title'], string $dynamicColumn = 'values'): self
+    public static function makeDefault(array $columns = [], array $dynamicKeys = ['title'], string $dynamicColumn = 'values', bool $strictSearch = false): self
     {
         return static::make('title')
             ->label('Titel')
             ->placeholder('Zoek op titel')
             ->description('Zoek op pagina titel')
-            ->query(FilterPresets::searchQuery($columns, $dynamicKeys, $dynamicColumn));
+            ->query(FilterPresets::searchQuery($columns, $dynamicKeys, $dynamicColumn, $strictSearch));
     }
 }